+++ States/recovered.py
@@ -1,0 +1,5 @@
+from state import State
+
+
+class Recovered(State):
+    pass
